Use `size_t` with name `buflen` consistently
A few of the parameters were named `num_read` "leaking" the use it was named
for in a calling function.  We use `buflen` with `buf` more consistently, and
use `size_t` were possible since we don't need or want to pass a negative value
for the size of a buffer.

+++ src/conn_sock.h
@@ -53,7 +53,7 @@ char *setup_seccomp_socket(const char *socket);
 char *setup_attach_socket(void);
 void setup_notify_socket(char *);
 void schedule_main_stdin_write();
-void write_back_to_remote_consoles(stdpipe_t pipe, char *buf, int len);
+void write_back_to_remote_consoles(stdpipe_t pipe, char *buf, size_t len);
 void close_all_readers();
 
 #endif // CONN_SOCK_H
